refactor: BookingActivity 내부 view, presenter, contract 패키지 제거
