+++ app/src/main/java/woowacourse/movie/ui/booking/BookingContract.kt
@@ -1,11 +1,11 @@
-package woowacourse.movie.ui.booking.contract
+package woowacourse.movie.ui.booking
 
+import java.time.LocalDate
+import java.time.LocalTime
 import woowacourse.movie.domain.model.Headcount
 import woowacourse.movie.domain.model.Movie
 import woowacourse.movie.domain.model.MovieSchedule
 import woowacourse.movie.domain.model.Theater
-import java.time.LocalDate
-import java.time.LocalTime
 
 interface BookingContract {
     interface Presenter {
