+++ app/src/main/java/woowacourse/movie/ui/booking/BookingPresenter.kt
@@ -1,15 +1,14 @@
-package woowacourse.movie.ui.booking.presenter
+package woowacourse.movie.ui.booking
 
+import java.time.LocalDate
+import java.time.LocalDateTime
+import java.time.LocalTime
 import woowacourse.movie.domain.model.Headcount
 import woowacourse.movie.domain.model.Movie
 import woowacourse.movie.domain.model.Theater
 import woowacourse.movie.domain.model.TheaterSchedules
 import woowacourse.movie.sample.DUMMY_MOVIES
 import woowacourse.movie.sample.DUMMY_THEATERS
-import woowacourse.movie.ui.booking.contract.BookingContract
-import java.time.LocalDate
-import java.time.LocalDateTime
-import java.time.LocalTime
 
 class BookingPresenter(
     private val bookingView: BookingContract.View,
