+++ app/src/main/java/woowacourse/movie/ui/booking/spinner/ScreeningTimeSpinner.kt
@@ -1,20 +1,19 @@
-package woowacourse.movie.ui.booking.view.spinner
+package woowacourse.movie.ui.booking.spinner
 
 import android.widget.AdapterView
 import android.widget.ArrayAdapter
 import android.widget.Spinner
-import java.time.LocalDate
+import java.time.LocalTime
 
-class ScreeningDateSpinner(
+class ScreeningTimeSpinner(
     private val spinner: Spinner,
-    dates: List<LocalDate>,
 ) {
-    init {
+    fun updateAdapter(times: List<LocalTime>) {
         spinner.adapter =
             ArrayAdapter(
                 spinner.context,
                 android.R.layout.simple_spinner_item,
-                dates,
+                times,
             )
     }
