+++ app/src/main/res/layout/activity_booking.xml
@@ -22,7 +22,7 @@
         android:id="@+id/main"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.booking.view.BookingActivity">
+        tools:context=".ui.booking.BookingActivity">
 
         <ScrollView
             android:id="@+id/layout_root_scroll"
